add active theme context
